new frame and compositor timing APIs for Platform (#9354)
* Add Compositor Timing APIs to Platform

These APIs are currently modeled after Android capabilities, but a lot of
the concepts are generic enough to map to other system with a compositor.

This adds Platfrom::queryCompositorTiming() which is currently only
implementd by PlaftormEGLAndroid. Vulkan and WebGPU will come later.
The functionality is implement via EGL, but a new NativeWindow
helper is provided to access the same data from non-EGL backends.
Also to be on the safe side, PlatformEGLAndroid reverts to using
the ANativeWindow if EGL fails.

This CL also cleans-up PlatformEGLAndroid to accomodate those changes.

* Add queryFrameTimestamps() API to Platform.

This API allows to query very precise timing information of
past frames on a given swapchain. This information is provided
by the Android framework via EGL or ANativeWindow (albeit
through private APIs).
Currently only implemented by PlatformEGLAndroid.